remove useAddAssetsActions hook
